+++ demo/application/src/main/java/com/gambler/StartApplication.java
@@ -1,0 +1,13 @@
+package com.gambler;
+
+import io.mongock.runner.springboot.EnableMongock;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+@EnableMongock
+public class StartApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(StartApplication.class, args);
+    }
+}
