+++ demo/application/src/main/java/com/gambler/config/ApplicationConfig.java
@@ -1,0 +1,16 @@
+package com.gambler.config;
+
+import gmongo.repository.UserRepository;
+import gmongo.repository.UserRepositoryImpl;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.mongodb.core.MongoTemplate;
+
+@Configuration
+public class ApplicationConfig {
+
+    @Bean
+    public UserRepository userRepository(MongoTemplate mongoTemplate) {
+        return new UserRepositoryImpl(mongoTemplate);
+    }
+}
