+++ demo/application/src/main/java/com/gambler/controller/TestController.java
@@ -1,0 +1,16 @@
+package com.gambler.controller;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class TestController {
+
+    @GetMapping("/test")
+    public String test() {
+        return "Hello Spring!";
+    }
+
+
+
+}
