+++ demo/application/src/main/java/com/gambler/error/exception/BackendException.java
@@ -1,0 +1,13 @@
+package com.gambler.error.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+/**
+ * Ошибка которую будем кидать сами с нашего бека
+ */
+@Getter
+@AllArgsConstructor
+public class BackendException extends RuntimeException {
+    private CabinetError error;
+}
