+++ demo/application/src/main/java/com/gambler/error/exception/ExceptionType.java
@@ -1,0 +1,19 @@
+package com.gambler.error.exception;
+
+/**
+ * Типы ошибок
+ */
+public enum ExceptionType {
+    /**
+     * Ошибка по нашей вине
+     */
+    LK,
+    /**
+     * Ошибка во внешней системе
+     */
+    EIS,
+    /**
+     * Логическая ошибка фронтов
+     */
+    LOGIC;
+}
