+++ demo/application/src/main/java/com/gambler/service/UserService.java
@@ -1,0 +1,10 @@
+package com.gambler.service;
+
+
+import gmongo.model.UserMng;
+
+import java.util.Optional;
+
+public interface UserService {
+    Optional<UserMng> findById(Integer id);
+}
