+++ demo/settings.gradle
@@ -1,1 +1,11 @@
+pluginManagement {
+    repositories {
+        gradlePluginPortal()
+    }
+}
+
 rootProject.name = 'demo'
+
+include 'gambler-mongo'
+include 'application'
+
