--- demo/src/main/java/com/example/demo/controller/TestController.java
@@ -1,16 +1,0 @@
-package com.example.demo.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class TestController {
-
-    @GetMapping("/test")
-    public String test() {
-        return "Hello Spring!";
-    }
-
-
-
-}
