--- demo/src/main/java/com/example/demo/service/UserService.java
@@ -1,9 +1,0 @@
-package com.example.demo.service;
-
-import com.example.demo.model.UserMng;
-
-import java.util.Optional;
-
-public interface UserService {
-    Optional<UserMng> findById(Integer id);
-}
