--- demo/src/main/java/com/example/error/exception/ExceptionType.java
@@ -1,19 +1,0 @@
-package com.example.error.exception;
-
-/**
- * Типы ошибок
- */
-public enum ExceptionType {
-    /**
-     * Ошибка по нашей вине
-     */
-    LK,
-    /**
-     * Ошибка во внешней системе
-     */
-    EIS,
-    /**
-     * Логическая ошибка фронтов
-     */
-    LOGIC;
-}
