--- demo/src/test/java/com/example/demo/ControllerTests.java
@@ -1,13 +1,0 @@
-package com.example.demo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ControllerTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
